fix(future): Resolve deployment issue by capitalizing component name
Correct the export function name in from page to Page to adhere to React component naming conventions. New to nextjs, so I forgot the page exported in nextjs is a react component
This change addresses the deployment failure caused by the function not being recognized as a React component due to its lowercase initial letter. The root cause of this issue was an oversight in the local build process, where the unsuccessful build went unnoticed.

This commit ensures that the Page component is correctly recognized during the build and deployment processes, preventing similar issues in future deployments.